lots of refactors
