+++ src/app/(auth)/404.tsx
@@ -1,4 +1,4 @@
-import NotFound from '../components/NotFound';
+import NotFound from '../../components/ui/NotFound';
 import { Button } from '@/components/ui/button';
 
 export default function Auth404Page() {
