+++ src/app/(auth)/application/application_components/PageStatus/ApplicationPageIndicator.module.css
