+++ src/app/(auth)/application/application_components/types.ts
