+++ src/app/qr/checkin_components/CheckInButton.tsx
