+++ src/app/qr/checkin_components/QRTicket.tsx
