+++ src/app/qr/checkin_components/SelectMeal.tsx
