+++ src/app/qr/checkin_components/UserNotFound.tsx
