+++ src/app/qr/meal/[mealType]/page.tsx
@@ -1,4 +1,4 @@
-import ScanPage from '@/app/qr/components/ScanPage';
+import ScanPage from '@/app/qr/checkin_components/ScanPage';
 
 type MealScanProps = {
     params: Promise<{ mealType: string }>;
